refactor users seed
